[Spring 지하철 노선도 관리 - 3단계] 소롱(최솔지) 미션 제출합니다.

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,39 +1,55 @@
 package wooteco.subway.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.exceptions.StationDuplicationException;
-import wooteco.subway.exceptions.StationNotFoundException;
-import wooteco.subway.repository.StationDao;
 import wooteco.subway.domain.station.Station;
+import wooteco.subway.dto.StationRequest;
+import wooteco.subway.dto.StationResponse;
+import wooteco.subway.exception.notRemovableException.NotRemovableStationException;
+import wooteco.subway.exception.duplicateException.StationDuplicationException;
+import wooteco.subway.exception.notFoundException.StationNotFoundException;
+import wooteco.subway.repository.SectionDao;
+import wooteco.subway.repository.StationDao;
 
 @Service
 public class StationService {
 
+    private static final int SIZE_OF_SECTION = 0;
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
-    public Station createStation(String name) {
-        validateDuplication(name);
-        Station station = new Station(name);
+    public StationResponse createStation(StationRequest stationRequest) {
+        Station station = stationRequest.toStation();
+        validateDuplication(station.getName());
+
         long id = stationDao.save(station);
         station.setId(id);
-        return station;
+        return StationResponse.of(station);
     }
 
-    private Station findById(long id) {
-        return stationDao.findById(id)
+    public StationResponse findById(long id) {
+        Station station = stationDao.findById(id)
             .orElseThrow(StationNotFoundException::new);
+        return StationResponse.of(station);
     }
 
-    public List<Station> findAll() {
-        return stationDao.findAll();
+    public List<StationResponse> findAll() {
+        List<Station> stations = stationDao.findAll();
+        return stations.stream()
+            .map(StationResponse::of)
+            .collect(Collectors.toList());
     }
 
     public void deleteById(Long id) {
+        if (sectionDao.findByStation(id)) {
+            throw new NotRemovableStationException();
+        }
         stationDao.deleteById(id);
     }
